requirements.txt
Flask
torch
transformers
pillow
git+https://github.com/huggingface/parler-tts.git
